feat(data): textiles y tope de recursos 32 -> 64 (1.9C) + cuatro temas del corpus
--- El tope, y por que se subio ---

El Director decidio anadir textiles. Al ir a hacerlo aparecio que quedaban DOS
huecos de recurso (30 de 32) y que el 32 NO era un numero redondo: era el ancho
de `dropoff_mask`, un uint32 con UN BIT POR RECURSO. Los textiles bien hechos
piden cinco o seis, porque lino egipcio, seda china, lana andina y algodon
indio son justamente lo que diferencia a unas civilizaciones de otras;
colapsarlos a 'fibra generica' habria borrado esa diferencia.

Se le planteo con el coste real y eligio subir a 64: dropoff_mask a uint64,
RESOURCE_COUNT a 64, save 16->17 y checksum 11->12. Es barato en memoria —unos
pocos KB— y se hace UNA vez y con margen, no cada vez que aparezca un material.
Quedan 28 huecos para lo que sugieran los 18 temas del corpus.

--- Los seis textiles ---

lana (1), lino (2), algodon (3), seda (5), tejido (2, producido) y fibra
sintetica (13, producida). Familia `textiles` nueva, anadida APPEND-ONLY al
enum para no renumerar las familias ya grabadas en el blob y en el HUD.

El algodon se domestico de forma INDEPENDIENTE en el valle del Indo y en
Mesoamerica, asi que sirve a las dos lineas sin inventar un prestamo cultural
que no existio. Todas las epocas van marcadas como PENDIENTES de contrastar con
el extracto de textiles del corpus: son diseño, no historia verificada todavia.

--- Un aviso que deja este cambio ---

Los indices de recurso se asignan alfabeticamente, asi que los seis textiles se
INTERCALARON y renumeraron todo lo que iba detras. Es correcto y determinista,
pero implica que anadir un recurso SIEMPRE obliga a subir save y checksum.
Anotado en el test del catalogo.

Y una equivocacion mia por el camino: al actualizar el test cambie dos valores
que parecian indices y eran EPOCAS —steel 12, uranium 14—. Los devolvi. Un sed
que empareja numeros sin leer el nombre de la estructura es exactamente asi de
peligroso.

--- Corpus: cuatro temas mas ---

construccion_hidraulica (40 [V], 216 citas, 20 obras), energia (50 [V], 258
citas), mineria (48 [V], 107 citas) y quimica (38 [V], 164 citas). Todos con
fuentes del Internet Archive y con sus NO VERIFICADO declarados.

Suite 43/43. Trayectorias intactas —1227, 1108, 10473—: los textiles todavia no
los consume nada.

+++ addons/chunsa_sim/core/include/chunsa/step.hpp
@@ -1627,7 +1627,7 @@ inline bool find_building_dropoff(const GameState& g, uint8_t owner, uint8_t res
     for (uint32_t j = 0; j < t.capacity; ++j) {
         if (!is_complete_owned_building(g, j, owner)) continue;
         const BuildingDefinitionV1& def = g.catalog->buildings[g.building_id[j]];
-        if ((def.dropoff_mask & (uint32_t{1} << resource_idx)) == 0u) continue;
+        if ((def.dropoff_mask & (uint64_t{1} << resource_idx)) == 0u) continue;
 
         FatalReason local_fatal = FatalReason::NONE;  // descartado a propósito, mismo patrón que combat/aggro
         const Vec2Fx there{Fx{g.pos_x[j]}, Fx{g.pos_y[j]}};
